requirements.txt
GitPython==2.0.3
Pillow==3.2.0
